+++ frontend/src/todo/TodoList.test.tsx
@@ -1,68 +1,25 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
-import TodoList, { Todo, TodoProps } from './TodoList';
+import { render } from '@testing-library/react';
+import TodoList from './TodoList';
+import { Todo } from './type';
 
-describe('<Todolist/>', () => {
-  const sampleTodo: Todo = {
-    id: '1',
-    text: 'TDD training',
-    done: false,
-  };
-  const setup = ({ todo = sampleTodo, onToggle, onRemove }: TodoProps) => {
-    const utils = render(
-      <TodoList todo={todo} onToggle={onToggle} onRemove={onRemove} />,
-    );
-    const { getByText } = utils;
-    const textSpan = getByText(todo.text);
-    const removeButton = getByText('×');
-    return {
-      ...utils,
-      textSpan,
-      removeButton,
-    };
-  };
-  const onToggle = jest.fn();
-  const onRemove = jest.fn();
+describe('<TodoList/>', () => {
+  const sampleTodolist: Todo[] = [
+    {
+      id: '1',
+      text: 'TDD training',
+      done: false,
+    },
+    {
+      id: '2',
+      text: 'going to home',
+      done: true,
+    },
+  ];
 
-  it('has span and button', () => {
-    const { textSpan, removeButton } = setup({
-      todo: sampleTodo,
-      onToggle,
-      onRemove,
-    });
-    expect(textSpan).toBeTruthy();
-    expect(removeButton).toBeTruthy();
-  });
-
-  it('shows line-through when done is true ', () => {
-    const { textSpan } = setup({
-      todo: { ...sampleTodo, done: true },
-      onToggle,
-      onRemove,
-    });
-    expect(textSpan).toHaveStyle('text-decoration: line-through;');
-  });
-
-  it('does not show line-through when done is false ', () => {
-    const { textSpan } = setup({
-      todo: { ...sampleTodo, done: false },
-      onToggle,
-      onRemove,
-    });
-    expect(textSpan).not.toHaveStyle('text-decoration: line-through;');
-  });
-
-  it('calls onToggle', () => {
-    const onToggle = jest.fn();
-    const { textSpan } = setup({ todo: sampleTodo, onToggle, onRemove });
-    fireEvent.click(textSpan);
-    expect(onToggle).toBeCalledWith(sampleTodo.id);
-  });
-
-  it('calls onRemove', () => {
-    const onRemove = jest.fn();
-    const { removeButton } = setup({ todo: sampleTodo, onToggle, onRemove });
-    fireEvent.click(removeButton);
-    expect(onRemove).toBeCalledWith(sampleTodo.id);
+  it('renders TodoList properly', () => {
+    const { getByText } = render(<TodoList todoList={sampleTodolist} />);
+    getByText(sampleTodolist[0].text);
+    getByText(sampleTodolist[1].text);
   });
 });
