+++ frontend/src/todo/type.tsx
@@ -1,0 +1,5 @@
+export interface Todo {
+  id: string;
+  text: string;
+  done: boolean;
+}
